update gui
started writing main code with gui functions, needs to be fixes as nothing actually work. need to figure out whatis wrong with 'createWindow" function and how to put a button in it (this  worked without classes). Also, main problem seems to appear when everything is put in functions, how else to create specific windows ?